Addition of queryableSecurable to eliminate circular ref, updates to add/delete/update in lists
