+++ src/sharepoint/rest/Queryable.ts
@@ -2,7 +2,6 @@
 
 import * as Util from "../../utils/util";
 import { Dictionary } from "../../collections/collections";
-import { RoleAssignments } from "./roleassignments";
 import { HttpClient } from "../../net/HttpClient";
 
 /**
@@ -67,6 +66,14 @@ export class Queryable {
     }
 
     /**
+     * Gets the parent url used when creating this instance
+     * 
+     */
+    protected get parentUrl(): string {
+        return this._parentUrl;
+    }
+
+    /**
      * Provides access to the query builder for this url
      * 
      */
@@ -112,7 +119,7 @@ export class Queryable {
      */
     public get(parser: (r: Response) => Promise<any> = (r) => r.json()): Promise<any> {
         let client = new HttpClient();
-        return client.get(this.toUrlAndQuery()).then(function(response) {
+        return client.get(this.toUrlAndQuery()).then(function (response) {
 
             if (!response.ok) {
                 throw "Error making GET request: " + response.statusText;
@@ -120,7 +127,7 @@ export class Queryable {
 
             return parser(response);
 
-        }).then(function(parsed) {
+        }).then(function (parsed) {
             return parsed.hasOwnProperty("d") ? parsed.d.hasOwnProperty("results") ? parsed.d.results : parsed.d : parsed;
         });
     }
@@ -129,7 +136,7 @@ export class Queryable {
 
         let client = new HttpClient();
 
-        return client.post(this.toUrlAndQuery(), postOptions).then(function(response) {
+        return client.post(this.toUrlAndQuery(), postOptions).then(function (response) {
 
             // 200 = OK (delete)
             // 201 = Created (create)
@@ -149,7 +156,7 @@ export class Queryable {
             // pipe our parsed content
             return parser(response);
 
-        }).then(function(parsed) {
+        }).then(function (parsed) {
 
             // try and return the "data" portion of the response
             return parsed.hasOwnProperty("d") ? parsed.d.hasOwnProperty("results") ? parsed.d.results : parsed.d : parsed;
@@ -194,78 +201,5 @@ export class QueryableInstance extends Queryable {
     public select(...selects: string[]): any {
         this._query.add("$select", selects.join(","));
         return this;
-    }
-}
-
-export class QueryableSecurable extends QueryableInstance {
-
-    /**
-     * Gets the set of role assignments for this item
-     * 
-     */
-    public get roleAssignments(): RoleAssignments {
-        return new RoleAssignments(this);
-    }
-
-    /**
-     * Gets the closest securable up the security hierarchy whose permissions are applied to this list item
-     * 
-     */
-    public get firstUniqueAncestorSecurableObject(): QueryableInstance {
-        this.append("FirstUniqueAncestorSecurableObject");
-        return new QueryableInstance(this);
-    }
-
-    /**
-     * Gets the effective permissions for the user supplied
-     * 
-     * @param loginName The claims username for the user (ex: i:0#.f|membership|user@domain.com)
-     */
-    public getUserEffectivePermissions(loginName: string): Queryable {
-        this.append("getUserEffectivePermissions(@user)");
-        this._query.add("@user", "'" + encodeURIComponent(loginName) + "'");
-        return new Queryable(this);
-    }
-
-    /**
-     * Breaks the security inheritance at this level optinally copying permissions and clearing subscopes
-     * 
-     * @param copyRoleAssignments If true the permissions are copied from the current parent scope
-     * @param clearSubscopes Optional. true to make all child securable objects inherit role assignments from the current object
-     */
-    public breakRoleInheritance(copyRoleAssignments = false, clearSubscopes = false): Promise<any> {
-
-        class Breaker extends Queryable {
-            constructor(baseUrl: string | Queryable, copy: boolean, clear: boolean) {
-                super(baseUrl, `breakroleinheritance(copyroleassignments=${copy}, clearsubscopes=${clear})`);
-            }
-
-            public break(): Promise<any> {
-                return this.post();
-            }
-        }
-
-        let b = new Breaker(this, copyRoleAssignments, clearSubscopes);
-        return b.break();
-    }
-
-    /**
-     * Breaks the security inheritance at this level optinally copying permissions and clearing subscopes
-     * 
-     */
-    public resetRoleInheritance(): Promise<any> {
-
-        class Resetter extends Queryable {
-            constructor(baseUrl: string | Queryable) {
-                super(baseUrl, "resetroleinheritance");
-            }
-
-            public reset(): Promise<any> {
-                return this.post();
-            }
-        }
-
-        let r = new Resetter(this);
-        return r.reset();
     }
 }
